Correct collectd cpu and memory plugin exceptions
The collectd cpu and memory plugins were failing to initialize properly
when the kubepods/k8s-infra cgroups were not configured.
The k8s-infra directory is present only after kubernetes is configured,
and only on kubernetes nodes.

This updates the collectd cpu and memory plugins to skip walking the
kubepods/k8s-infra directory tree if that path does not exist.
The summary algorithm works correctly when skipping this path since
it does not need to add values that do not exist. The k8s-system and
k8s-addon values will be reported as 0.

This also removes the benign error and warning logs that came out
when there is no worker_reserved.conf file, i.e., non-worker nodes.

Change-Id: I36ecd689dad45543f5a7cc6c7bc99a21c00c8122
Closes-Bug: 1849511
Signed-off-by: Jim Gauld <james.gauld@windriver.com>