Add ptp no-lock alarm support for controllers
The current PTP No-Lock alarm is not raised for
controllers that assume Grandmastership.

This is incorrect behavior. Every node configured
for PTP needs to be locked to a remote Grandmaster.

This update increases the scope of the no lock alarm
to apply to all node types ; computes, storage and
now controllers as well.

Test Plan:

PASS: Verify PTP no lock alarm for controller.

Change-Id: I271c8ad235e6cb680bdb27f377149b7e1a3d14cd
Closes-Bug: 1850554
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>