Fix NTP alarm bug with IP addresses
This submit fixes a bug that would prevent NTP alarms to be raised on
IP addresses less than 10 characters. Updates the header documentation
and cleans up return codes for _raise_alarm(). The submit also changes
the audit time from 10 minutes to 5 minutes so that alarms can be
cleared faster and also align with the PTP audit

Change-Id: I07797956be6db94684cf4d10960ebb43f703f32e
Closes-bug: 1851599
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>