Added retry mechanism to clear port alarm
When the link state transitions from DOWN to UP, the current collectd
process attempts to clear the alarm once and once only. If such attempt
failed, no further attempts will be made, and the alarm will persist in
fm alarm-list.

This fix added an additional check to ensure that as long as port alarm
is not cleared and the link state is UP, it will attempt to clear the
port alarm.

Closes-Bug: 1871453

Change-Id: Iaa65f64808272a5760e655a33c14810df51e28b1
Signed-off-by: Thomas Gao <Thomas.Gao@windriver.com>