Format 2 lines ntpq data into 1 lines
The problem was logic expected one line data for
ntpq result. But it was 2 lines for each ntp server
entry. When peer server is selected, script checked
refid if refid is reliable or not but it could not
find because refid is in the following line.
This fix formats 2 lines data into 1 line.

The minor alarm "minor alarm "NTP cannot reach
external time source; syncing with peer controller
only" is removed because NTP does not prioritize
external time source over peer.

Closes-Bug: 1889101

Signed-off-by: Takamasa Takenaka <takamasa.takenaka@windriver.com>
Change-Id: Icc8316bb1a7041bf0351165c671ebf35b97fa3bc