[PTP SyncE] Handle disabled synce
When synce is disabled, the clock config file would be empty.
This commit is to handle the scenario where synce is disabled.

This update also improves the way in extracting the instance name
from the ptp4l config filename.

Story: 2009130
Task: 44542

Test Plan:
Pass: Test that clock is not monitored when synce is disabled.

Change-Id: Iec8bdeeb1fc7d94cdb6835e4f347e618d6c463d1
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>