[PTP SyncE] Determine primary nic from tty port
The interface in the ts2phc config file may not be the same
NIC as the tty port. It should not be referenced to determine
the primary NIC.
This commit is to find the interface name from the tty pci address
and disregard the interface name in the config file.

Story: 2009130
Task: 44609

Test Plan:
Pass: Check for GNSS alarm when ts2phc tty is set to a NIC
      without GNSS signal

Signed-off-by: Teresa Ho <teresa.ho@windriver.com>
Change-Id: Ie0d12cbd1443f5e18801be099543d4fdc63942e6