[PTP SyncE] Fix clockClass and timeTraceable
The port specified in the ptp4l conf file can be of any of the ports
in the NIC. To determine whether the ptp4l is associated with the
primary NIC, it needs to convert the port into the base port.

This commit also corrected the timeTraceable setting for the
holdover cases.

Story: 2009130
Task: 44701

Test Plan:
Pass: Assign a non-base port to the ptp4l instance.
      Set locked, holdover or freerun conditions.
      Ensure that the clockClass and timeTraceable are set correctly

Change-Id: I3cf2c41ee4a90dc2b09d37b2f2e5071e03ba2fd5
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>