[PTP SyncE] Raise alarm if no pin configured in clock instance
If no pin is configured for the secondary NIC in the clock
instance, the status was not reported. This commit is to raise an
alarm when there is missing configuration.

Story: 2009130
Task: 44744

Test Plan:
Pass: Configure clock instance with no ptp parameter for pins
      Check that an alarm is raised
Pass: Add the ptp parameter for the pin for the clock interface
      Check that the alarm is cleared

Change-Id: Icd8e7ea7511dcc3ae9c5771c3b46707c043e9849
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>