Fixing PTP conf files in Debian for alarming
In Debian the PTP *.conf files moved from /etc/ptpinstance/ to
/etc/linuxptp/ptpinstance, and therefore collectd plugin would not be
able to load the configuration and could not generate the proper alarms.
This change sets the path of configuration files according to the OS
release ('centos' or 'debian') as read from /etc/os-release.
As a bonus, it also fixes the calls to subprocess.check_output() to
decode() the result, which in Python3 is of type "bytes", and that could
generate a type exception when used with string functions like split().

Test Plan:
PASS: Checked alarm generation with PTP configuration and clock skew.

Closes-Bug: 1978576
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: I049eb4f1841c2fe0d07bdf3a337d53eea36f783d