Fix to prevent truncating IPv6 value when NTP alarm is triggered.
When the NTP alarm is triggered indicating that the peer is not
reachable and if it is IPv6, the IP value is truncated.
This occurs because the NTP plugin relies on the output of the
ntpq -np command, which shows the truncated IPv6 as well.
This causes the IPv6 in the alarm to be truncated, showing its
partial information.
To fix this, a mechanism was implemented to invoke the ntpq
command but specifying the association corresponding to the IPv6
which is truncated. In this way, detailed information of the
association is retrieved, including the full IPv6.
That IPv6 will be the one that will be used as the value for
the alarm.

Closes-Bug: 2004043

Test Plan:

PASS: Configure NTP with unreachable IPv6 peers to trigger the
NTP alarm
PASS: Configure NTP with reachable IPv6 peers to avoid alarms
PASS: Configure NTP with unreachable IPv4 peers to trigger the
NTP alarm
PASS: Configure NTP with reachable IPv4 peers to avoid alarms

Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>
Change-Id: Id7e0af4f130f04c5eb037e5ff0d0a0cc5ce71b3e