Update vm-topology debian package ver based on git
Update debian package versions to use git commits for:
 - vm-topology

Old version was: 1
New version is: 15

The Debian packaging has been changed to reflect all the
git commits under the directory, and not just the commits
to the metadata folder.

This ensures that any new code submissions under those
directories will increment the versions.

Note:  vm-topology is not currently setup to build or
install on Debian.
 This is because it requires libvirt python components to
run which are not installed on the platform.
 In order to build this package, the debian_pkg_dirs file
needed to be temporarily updated to include vm-topology

Test Plan:
  PASS: build-pkgs -p vm-topology

Story: 2010550
Task: 47410
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I8aa233a6c3daaa68f0a2ad1af33365d320a68665