Add the NFD namespace to the platform infrastructure list in collectd
Actually, the pods installed in the node-feature-discovery namespace run
on application cores. Although, the Node Feature Discovery App is seen
as a platform app, and therefore, its pods need to run on platform
cores.

So, in this change, the node-feature-discovery namespace will be added
in the collectd platform infrastructure list to follow the changes did
in kubelet

Test Plan:
PASS: Execute kube-memory tool and ensure its output contains nfd namespace data
PASS: Edit collectd tool source code to display the list of namespaces classified as K8S_NAMESPACE_SYSTEM and verify if the node-feature-discovery namespace was found.

Story: 2010769
Task: 48326

Depends-On: https://review.opendev.org/c/starlingx/integ/+/887743

Change-Id: Idc31c36fc4e12d4f91b2fa45ed9dee663cb024f9
Signed-off-by: Marcos Paulo Oliveira Silva <Marcos.PauloOliveiraSilva@windriver.com>