Enhance ptp alarming logic
Fix 1:
As per the HA phc2sys design, the query for checking the forced lock
state of HA phc2sys has been updated to 'forced lock'

Fix 2:
Correct the handling for setting the ptp4l clockClass when there is no
ts2phc instance configured to set time on that NIC.

Fix 3:
Upgrade the logic in process_ptp_synce(). NICs may be configured with
both a GNSS connection and an SMA connection, so it is necessary to
check for a GNSS lock first. If GNSS is locked, no further checking is
required. If GNSS is lost, check if SMA is configured and locked before
raising alarm.

Test plan:
Pass: Validate that 'forced lock' query works correctly
Pass: Validate that ptp4l clockClass remains 248 when there is no ts2phc
instance configured to set time on a NIC
Pass: Validate that a NIC with both GNSS and SMA connection configured
does not incorrectly raise an alarm if GNSS is present

Story: 2010723
Task: 48602

Change-Id: I92869c7c28c786c5e68cd493ae6582b9b4884c21
Signed-off-by: Cole Walker <cole.walker@windriver.com>