Generalize platform pods check using platform label
This commit introduces a check for platform label, which if
present, the cpu and memory for these pods will be
considered in platform resources. Earlier, pod was
considered as part of platform resources by hardcoded namespaces,
which are tiresome to maintain. Now pods are treated as 'platform'
if the namespace is in set of hard-coded namespaces, or if
the component label is 'app.starlingx.io/component=platform'.

Test Plan:
   Pass: Code changes on AIO-SX and collectd was restarted successfully
   Pass: Verify that /var/log/collectd.log contains cpu and memory
  consumption info for platform pods.

Story: 2010904
Task: 48718
Change-Id: Ia8442717009f92dbe022f9512e226913c45d9473
Signed-off-by: sshathee <shunmugam.shatheesh@windriver.com>