Update source selection alarm to 'msg' type
Change the alarm type for HA phc2sys source selection changes to type
'msg'. It is not necessary for this event to raise/clear an alarm, as
this warning is intended for informational purposes and does not require
intervention from the user. This allows for source selection changes to
be audited without being treated as a system fault.

Test plan:
PASS: Verify ptp plugin startup and general operation
PASS: Verify that source selection changes generate a log entry in 'fm
event-list'
PASS: Verify other HA phc2sys alarms are raised correctly and display in
fm alarm-list

Story: 2010723
Task: 49393

Change-Id: I7632e79ffd7522eb4e7caa34a1959aa62d3e8879
Signed-off-by: Cole Walker <cole.walker@windriver.com>