PTP: fix handling query_pmc response
Depending on the query_pmc response it could cause the PTP plugin
to crash and become unavailable for some time. For example, when
the ptp4l is unavailable or incapable to answer the pmc query.

Test Plan:
Pass: Verify ptp plugin startup and operation
Pass: Verify operation when restarting ptp service
Pass: Verify operation when ptp4l is down
Pass: verify operation when a ptp4l is manually started and
stopped.

Closes-Bug: 2068025

Change-Id: I7f54377443489669c019d352682ba57f583976fe
Signed-off-by: Andre Mauricio Zelak <andre.zelak@windriver.com>