PTP: Correct wording of collectd alarm when ptp4l instance is stopped
The collectd alarm text that is triggered when a ptp4l instance is
stopped indicates that the fault is due to the timestamping mode,
which is incorrect and potentially confusing for customers.

This text is updated to reflect the actual cause of the alarm.

TEST PLAN:
PASS: Create ptp4l instances and apply
      systemctl stop ptp4l@ptp-inst1
      check alarm to see as corrected text.
      systemctl stop ptp4l@ptp-inst2
      check alarm to see as corrected text.
      start both ptp4l, no alarms.
PASS: Run in hardware timestamping mode in VM
      check alarm

Story: 2011370
Task: 52051

Signed-off-by: Tara Nath Subedi <tara.subedi@windriver.com>
Change-Id: If36ad594846235521e9f00c9aeb3a20c5674a2ee