PTP: Stopping ts2phc service doesn't trigger switchover
PTP Redundant T-GM, with two ts2phc instance and gnss sources,
stopping currently selected ts2phc service doesn't trigger
switchover.

The collectd service is responsible for setting clock-class on ptp
device (SET GRANDMASTER_SETTINGS_NP) based upon DPLL status on CGU
(/sys/kernel/debug/ice/<PCI>/cgu) when the ptp device is part
ts2phc slave configuration. Clock class is being set 6, 7/140 or
248 depending upon DPLL status is locked, holdover or invalid.

It is not being checked whether the ts2phc service itself is
running or not, and clock class is solely dependent upon DPLL
status. Because of this when ts2phc service is down and it is not
updating offset, the clockclass still could be 6 or 7 and also
announcing this clockclass downstream. In HA configuration, as
the clockclass is still better, there won't be switchover.

This commit fixes the issue by considering whether the ts2phc
is running or not, when ts2phc is not running and the ptp device
is local GM (grandmasterIdentity and clockIdentity equal)  the
clock-class is set to 248 and time and frequency traceability
also set to false. In case of ts2phc is not running and the ptp
device is not local GM, clock-class is untouched, to let synced
with PTP source itself.

TEST PLAN:
PASS: Setup ts2phc service and stop ts2phc, then check
      PARENT_DATA_SET, GRANDMASTER_SETTINGS_NP to confirm
      clockclass changed from 6/7 to 248.
      Also check time/frequency traceability
      (GRANDMASTER_SETTINGS_NP and TIME_PROPERTIES_DATA_SET)
      changed from true to false.
      start ts2phc, confirm clockclass changed back to 6/7
      and time/frequency traceability changed to true.
PASS: Setup HA with two ts2phc instance and gnss sources,
      stopping currently selected ts2phc service trigger
      switchover.
PASS: Same ptp device with both ts2phc and ptp source
      stop ts2phc, ptp device's portstate changed to slave.
      check clock-class (PARENT_DATA_SET), and time/frequency
      traceability (TIME_PROPERTIES_DATA_SET), it should be
      as received on announce message.

Story: 2011370
Task: 52279

Change-Id: Ic5d0b5e62322e76d9218cfcb4626148978a760a5
Signed-off-by: Tara Nath Subedi <tara.subedi@windriver.com>