Improve alarms for stopped PTP services
Improve the alarms to explicitly indicate when a PTP service is enabled
but not running.

This change makes a tweak to the Phc2sys instance parsing to add an
explicit entry for phc2sys even when HA is not enabled. Additionally,
the check for service state of each configured PTP service now includes
phc2sys and ts2phc.

Test plan:
Pass: Verify collectd-extensions package build
Pass: Verify that alarms are raised and cleared when each PTP service is
stopped and started
Pass: Verify general alarm operation (Out of Threshold, GNSS signal
loss, Not Locked to GM)
Pass: Verify collectd logs are free of errors and that process state
logs are present

Story: 2011345
Task: 52365

Change-Id: I3daca025ba6005bb47030a9e2dac38c250aac01d
Signed-off-by: Cole Walker <cole.walker@windriver.com>