Correct logic for setting freerun clockClass
When multiple ts2phc instances are configured, collectd does not
correctly use the holdover_timestamp for the relevant instance because
the loop continues and the value is clobbered.

This change breaks out of the loop after the holdover_timestamp is found
so that it does not get clobbered by further processing.

Additionally, the EEC and PPS pins of a DPLL device can reach a split
state where one reports holdover and the other freerun. Update the logic
to check both pins at all times and treat the worse condition as the
state for that DPLL.

Test plan:
Pass: Verify that holdover->freerun transiton works properly with one
and two ts2phc instances.
Pass: Verify that appropriate clockClass is set
Pass: Verify that holdover delta is logged
Pass: Verify that freerun is reached when one or more pin sources are
degraded
 - Repeatedly toggle GNSS state until the EEC and PPS pins are in a
   mixed state (holdover+freerun) and verify that collectd treats the
   "worst" condition as the current state.

Story: 2011345
Task: 52784

Change-Id: Id5ca79d762a72094a3d6d8019cc2449390953191
Signed-off-by: Cole Walker <cole.walker@windriver.com>