PTP plugin: Fix not locked to remote GM alarm in T-BC setups
In a system using the T-BC configuration with two or more PTP4L
instances, the instance that communicates with the grandmaster
is called upstream instance, and the instances that communicates
with ordinary clocks are called downstream.

When the upstream instance was still connnected to an external
GM but its clock class degraded, this patch prevents the alarm
"controller-0 is not locked to remote PTP Grand Master" from
being triggered on downstream instances. The logic now ensures
that the "no locked" alarm is raised only when the communication
to the grandmaster fails.

Test Cases:

PASS: Configure a system as PTP T-BC, with at least two PTP4L
      instances, one as upstream, and another as downstream.

      Configure the phc2sys instance to synchronize the upstream
      PHC to the system clock:

      system ptp-interface-add phc0if1 phcinst0
      system host-if-ptp-assign controller-0 oam0 phc0if1

      And configure the ts2phc instance to synchronize the system
      clock to the downstream PTP4L PHC:

      system ptp-instance-add tsinst0 ts2phc
      system host-ptp-instance-assign controller-0 tsinst0
      system ptp-instance-parameter-add tsinst0
          cmdline_opts="-s generic"
      system ptp-interface-add ts0if1 tsinst0
      system host-if-ptp-assign controller-0 ptp02 ts0if1
      system ptp-interface-parameter-add ts0if1 ts2phc.channel=1
      system ptp-interface-parameter-add ts0if1 ts2phc.pin_index=1

      Verify no alarms are raised for none of the PTP4L instances.

PASS: Disconnect the external Grand Master connected to upstream PTP4L
      instance, and verify the "controller-0 is not locked to remote
      PTP Grand Master" alarm is raised on both upstream, and
      downstream PTP4L instances.

PASS: Reconnected the external Grand Master connected to upstream
      PTP4L instance, and verify the alarms were cleared.

PASS: Downgrade the Grand Master clock class to 165, and verify there
      is no "controller-0 is not locked to remote PTP Grand Master"
      alarm raised.

Closes-bug: 2136223

Change-Id: Idaba470923745550bf4dda390e312345a5d17481
Signed-off-by: Andre Mauricio Zelak <andre.zelak@windriver.com>